Making PUT API reactive and storing memTable and indices in Sailee.java which is an Observable, which updates SegmentGenerator which is an Observer which will flush memTable to memory when it's full.

+++ src/main/java/com/mydb/mydb/MydbApplication.java
@@ -10,11 +10,11 @@ import java.io.File;
 @SpringBootApplication
 public class MydbApplication {
 
-  public static long MAX_MEM_TABLE_SIZE = 100;
+  public static int MAX_MEM_TABLE_SIZE = 100;
 
   public static void main(String[] args) {
     if (args.length > 0)
-      MAX_MEM_TABLE_SIZE = Long.parseLong(args[0]);
+      MAX_MEM_TABLE_SIZE = Integer.parseInt(args[0]);
     makeDirs();
     SpringApplication.run(MydbApplication.class, args);
   }
